Rename to template-default
