+++ lib/middleware/template-default.js
